requirements.txt
streamlit==1.28.1
firebase-admin==6.1.0
pandas==2.0.3
python-dotenv==1.0.0
twilio==8.5.0
sendgrid==6.10.0
pytz==2023.3
plotly==5.17.0
numpy==1.24.3
xlsxwriter==3.1.0
reportlab==4.0.4
PyJWT==2.8.0
phonenumbers==8.13.0
names==0.3.0
pytest==7.4.0